Don't load LDAP authentication module if no requirements are installed
The build fails, because no module for django_auth_ldap could
be found. To resolve this issue, we now require the user to
explicitly configure the AUTH_MECHANISM or use NodeAuthBackend
as default.
NodeAuthBackend is also the fallback if LDAP was configured
but the required packages were not installed. This way the
software will always run but the user can see within the logs
that some problem occured.