Renamed UserAuthenticationSource to AuthenticationSource
This way even groups can be imported form external authentication sources.
Furthermore this commit removed access to ui-elements depending on user permissions.
The permissions can be given not only user wise but also group wise. This way
it is possible to create an "administration" group which is able to administer
the application (managing users and groups)